requirements.txt
python-telegram-bot==13.4.1
sqlalchemy
apscheduler
mysql-connector-python
python-dotenv